pyproject.toml
[project]
name = "book-recommender-backend"
version = "0.1.0"
description = "backend for book recommendations"
authors = [
    {name = "rong2ren",email = "98350815+rong2ren@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.9,<4.0"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "supabase (>=2.14.0,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "sentence-transformers (>=4.0.1,<5.0.0)",
    "pydantic-settings (>=2.8.1,<3.0.0)",
    "loguru (>=0.7.3,<0.8.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
mypy = "^1.15.0"
pytest = "^8.3.5"
pre-commit = "^4.2.0"

[tool.black]
line-length = 88  # Standard PEP 8 width
target-version = ["py311"]  # Matches your Python version
include = '\.pyi?$'  # Format .py and .pyi files
exclude = '''
/(
    \.git
    | \.venv
    | build
    | dist
    | migrations
)/
'''

[tool.mypy]
python_version = "3.11"
warn_return_any = true
disallow_untyped_defs = true
plugins = "pydantic.mypy"
# Add these to support FastAPI better
ignore_missing_imports = true  # Avoid errors for optional deps
show_error_codes = true
strict_equality = true

[tool.pytest.ini_options]
python_files = "test_*.py"
addopts = "-v --cov=app --cov-report=term-missing"
asyncio_mode = "auto"

[tool.poetry]
name = "book-recommender-backend"
packages = [
  { include = "app" }  # Points to your code directory
]